+++ lib/pages/driver.ex
@@ -15,6 +15,9 @@ defmodule Pages.Driver do
   @doc "Click an element within a page. Implementation for `Pages.click/4`."
   @callback click(Pages.Driver.t(), Pages.http_method(), binary(), Hq.Css.selector()) :: Pages.Driver.t() | no_return()
 
+  @doc "Render a change. Implementation for `Pages.render_change/3`."
+  @callback render_change(Pages.Driver.t(), Hq.Css.selector(), Enum.t()) :: Pages.Driver.t()
+
   @doc "Re-renders the page. Implementation for `Pages.rerender/1`."
   @callback rerender(Pages.Driver.t()) :: Pages.Driver.t()
 
@@ -34,5 +37,13 @@ defmodule Pages.Driver do
   @callback with_child_component(Pages.Driver.t(), child_id :: binary(), (Pages.Driver.t() -> term())) ::
               Pages.Driver.t()
 
-  @optional_callbacks [click: 4, rerender: 1, submit_form: 2, submit_form: 4, update_form: 4, with_child_component: 3]
+  @optional_callbacks [
+    click: 4,
+    render_change: 3,
+    rerender: 1,
+    submit_form: 2,
+    submit_form: 4,
+    update_form: 4,
+    with_child_component: 3
+  ]
 end